backend/python/requirements.txt
fastapi
uvicorn
googletrans==4.0.0-rc1
gTTS
soundfile
sounddevice
requests
python-dotenv
scikit-learn
python-multipart
pandas
feedparser
httpx
